// src/generated/models/Cr133_announcementsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cr133_announcementscr133_status = {
  100000000: 'Draft',
  100000001: 'Test Sent',
  100000002: 'Sent'
} as const;
export type Cr133_announcementscr133_status = keyof typeof Cr133_announcementscr133_status;
export const Cr133_announcementsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cr133_announcementsstatecode = keyof typeof Cr133_announcementsstatecode;
export const Cr133_announcementsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cr133_announcementsstatuscode = keyof typeof Cr133_announcementsstatuscode;

export interface Cr133_announcementsBase {
  cr133_announcementid: string;
  cr133_bodyfieldsjson?: string;
  cr133_image?: string;
  cr133_name?: string;
  cr133_recipients?: string;
  cr133_renderedhtml?: string;
  cr133_sentat?: string;
  cr133_status?: Cr133_announcementscr133_status;
  cr133_subjectar?: string;
  cr133_subjecten?: string;
  "cr133_Template@odata.bind"?: string;
  cr133_testsentat?: string;
  cr133_testsentto?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cr133_announcementsstatecode;
  statuscode?: Cr133_announcementsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cr133_announcements extends Cr133_announcementsBase {
  cr133_image_timestamp?: number;
  cr133_image_url?: string;
  cr133_imageid?: string;
  cr133_statusname?: string;
  cr133_templatename?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  cr133_template?: object;
  _cr133_template_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}
export type Cr133_announcementsImageColumnName = 'cr133_image';

export type Cr133_announcementsUploadColumnName = Cr133_announcementsImageColumnName;